wip: server & database + comms

+++ crates/engine/src/window.rs
@@ -1,5 +1,5 @@
 #[derive(Copy, Clone, Eq, PartialEq, Hash)]
 pub struct Size {
-    pub width: u32,
-    pub height: u32,
+    pub width: u16,
+    pub height: u16,
 }
